code/requirements.txt
kafka-python-ng==2.2.3
prometheus-client==0.19.0
requests==2.31.0

# pandas==2.1.3
# numpy==1.26.2
# matplotlib==3.8.2
